analytics/requirements.txt
pymongo
mysql.connector
Flask
mysql-connector-python